backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anirudh Konduru <anirudhmkonduru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask-shell-ipython = "^0.4.1"
flask = "^2.2.2"
flask-cors = "^3.0.10"
ipykernel = "^6.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
